% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crucial_names.R
\name{check_crucial_names}
\alias{check_crucial_names}
\title{Check if a named object contains expected names}
\usage{
check_crucial_names(x, expected_names)
}
\arguments{
\item{x}{A named object.}

\item{expected_names}{String; expected names of \code{x}.}
}
\value{
Invisible \code{x}, or an error with informative message.
}
\description{
It is based on:
\url{https://www.rdocumentation.org/packages/fgeo.tool/versions/1.2.5/topics/check_crucial_names}.
}
\examples{
v <- c(name_a = 1)
check_crucial_names(v, "name_a")
try(check_crucial_names(v, "name_b"))

df <- data.frame(name_a = 1)
check_crucial_names(df, "name_a")
try(check_crucial_names(df, "name_b"))
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{degrees}()},
\code{\link{dropbox_exists}()},
\code{\link{path_examples}()},
\code{\link{r2dii_packages}()}
}
\concept{miscellaneous utility functions}
