% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_column_names.R
\name{clean_column_names}
\alias{clean_column_names}
\title{A version of \code{\link[janitor:clean_names]{janitor::clean_names()}} that also cleans dplyr groups}
\usage{
clean_column_names(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
A version of \code{\link[janitor:clean_names]{janitor::clean_names()}} that also cleans dplyr groups
}
\examples{
library(dplyr)

clean_column_names(group_by(tibble(x.x = 1), x.x))

# Cleans names but not groups
janitor::clean_names(group_by(tibble(x.x = 1), x.x))
}
\seealso{
\code{\link[janitor:clean_names]{janitor::clean_names()}}, \code{\link[=unclean_column_names]{unclean_column_names()}}.

Other functions to deal with unclean names: 
\code{\link{clean_quos}()},
\code{\link{clean_quo}()},
\code{\link{unclean_column_names}()}
}
\concept{functions to deal with unclean names}
