% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_names.R
\name{clean_quo}
\alias{clean_quo}
\title{Clean quo expression}
\usage{
clean_quo(quo)
}
\arguments{
\item{quo}{A quosure to test.}
}
\value{
A quosure.
}
\description{
Clean quo expression
}
\examples{
unclean_quo <- rlang::quo(x.x)
unclean_quo

clean_quo(unclean_quo)
}
\seealso{
Other functions to deal with unclean names: 
\code{\link{clean_column_names}()},
\code{\link{clean_quos}()},
\code{\link{unclean_column_names}()}
}
\concept{functions to deal with unclean names}
