% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_names.R
\name{clean_quos}
\alias{clean_quos}
\title{Clean a list of quosures}
\usage{
clean_quos(quos)
}
\arguments{
\item{quos}{A list of quosures}
}
\value{
A list of quosures.
}
\description{
Clean a list of quosures
}
\examples{
quos <- rlang::quos(x.x, y.y)
quos

clean_quos(quos)
}
\seealso{
Other functions to deal with unclean names: 
\code{\link{clean_column_names}()},
\code{\link{clean_quo}()},
\code{\link{unclean_column_names}()}
}
\concept{functions to deal with unclean names}
