% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create a configuration file}
\usage{
create_config(..., path = fs::file_temp("config-", ext = ".yml"))
}
\arguments{
\item{...}{String(s) to pass to the \code{...} argument of \code{\link[glue:glue]{glue::glue()}}.}

\item{path}{String giving the path where to save the configuration file.}
}
\value{
Invisible \code{path}.
}
\description{
Create a configuration file
}
\examples{
out <- create_config("
  default:
    field: value
")
out

show_config(out)
}
\seealso{
\link{FINANCIAL.TIMESTAMP}

Other functions to deal with configuration files: 
\code{\link{default_config}()},
\code{\link{edit_config}()},
\code{\link{show_config}()}
}
\concept{functions to deal with configuration files}
