% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbox_exists.R
\name{dropbox_exists}
\alias{dropbox_exists}
\title{Does this computer have a local copy of 2dii's dropbox folder?}
\usage{
dropbox_exists()
}
\value{
Logical.
}
\description{
Does this computer have a local copy of 2dii's dropbox folder?
}
\examples{
dropbox_exists()
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{check_crucial_names}()},
\code{\link{degrees}()},
\code{\link{path_examples}()},
\code{\link{r2dii_packages}()}
}
\concept{miscellaneous utility functions}
