% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_config.R
\name{edit_config}
\alias{edit_config}
\title{Edit the default configuration file}
\usage{
edit_config(path = get_config())
}
\arguments{
\item{path}{A string giving the path to a configuration file. \code{NULL} defaults
to \code{\link[=get_config]{get_config()}}.}
}
\value{
Invisible \code{default_config()}.
}
\description{
Edit the default configuration file
}
\examples{
# Use only interactively
# edit_config(example_config("config-toy.yml"))

# Now you should see something like this:
show_config(example_config("config-toy.yml"))
}
\seealso{
\link{FINANCIAL.TIMESTAMP}

Other functions to deal with configuration files: \code{\link{default_config}},
  \code{\link{show_config}}
}
\concept{functions to deal with configuration files}
