% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{FINANCIAL.TIMESTAMP}
\alias{FINANCIAL.TIMESTAMP}
\alias{timestamps_datastore_export}
\alias{ALD.TIMESTAMP}
\alias{DATASTORE.TIMESTAMP}
\alias{DATAPREP.TIMESTAMP}
\alias{START.YEAR}
\alias{TIME.HORIZON}
\alias{RISK.YEAR}
\alias{ADDITIONAL.YEAR}
\alias{TECH.LIST}
\alias{TECH.EXCLUDE}
\alias{SECTOR.LIST}
\alias{OTHER.SECTOR.LIST}
\alias{SCENARIO.SOURCES.LIST}
\alias{IEA.SCENARIO.LIST}
\alias{WEB.REGION.LIST}
\alias{SCENARIO.GEOGRAPHIES.LIST}
\alias{EQUITY.MARKET.LIST}
\alias{GLOBAL.AGGREGATE.SECTOR.LIST}
\alias{GLOBAL.AGGREGATE.SCENARIO.SOURCES.LIST}
\alias{CREATE.META.PORTFOLIO}
\alias{META.INVESTOR.NAME}
\alias{META.PORTFOLIO.NAME}
\alias{HasBV}
\alias{HasRISK}
\alias{HasMAP}
\alias{HasSB}
\title{Access parameters from a configuration file}
\usage{
FINANCIAL.TIMESTAMP(file = get_config())

timestamps_datastore_export(file = get_config())

FINANCIAL.TIMESTAMP(file = get_config())

ALD.TIMESTAMP(file = get_config())

DATASTORE.TIMESTAMP(file = get_config())

DATAPREP.TIMESTAMP(file = get_config())

START.YEAR(file = get_config())

TIME.HORIZON(file = get_config())

RISK.YEAR(file = get_config())

ADDITIONAL.YEAR(file = get_config())

TECH.LIST(file = get_config())

TECH.EXCLUDE(file = get_config())

SECTOR.LIST(file = get_config())

OTHER.SECTOR.LIST(file = get_config())

SCENARIO.SOURCES.LIST(file = get_config())

IEA.SCENARIO.LIST(file = get_config())

WEB.REGION.LIST(file = get_config())

SCENARIO.GEOGRAPHIES.LIST(file = get_config())

EQUITY.MARKET.LIST(file = get_config())

GLOBAL.AGGREGATE.SECTOR.LIST(file = get_config())

GLOBAL.AGGREGATE.SCENARIO.SOURCES.LIST(file = get_config())

CREATE.META.PORTFOLIO(file = get_config())

META.INVESTOR.NAME(file = get_config())

META.PORTFOLIO.NAME(file = get_config())

HasBV(file = get_config())

HasRISK(file = get_config())

HasMAP(file = get_config())

HasSB(file = get_config())
}
\arguments{
\item{file}{A string giving a path to a configuration file.}
}
\value{
The value of the requested field. It's base time depends on the
field.
}
\description{
Access parameters from a configuration file
}
\examples{
# Using the default configuration file that comes with the package
FINANCIAL.TIMESTAMP()

# Using your configuration file
path_to_your_config_file <- example_config("config-toy.yml")

# Passing your configuration file via the `file` argument
FINANCIAL.TIMESTAMP(file = path_to_your_config_file)

# The default remains the same
FINANCIAL.TIMESTAMP()
ALD.TIMESTAMP()

# Setting the path to your configuration file globally
old_options <- options(r2dii.dataraw_config = path_to_your_config_file)
# The default configuration file has now changed
FINANCIAL.TIMESTAMP()
ALD.TIMESTAMP()

# Reset options to cleanup
options(old_options)

timestamps_datastore_export()
FINANCIAL.TIMESTAMP()
DATASTORE.TIMESTAMP()
DATAPREP.TIMESTAMP()
START.YEAR()
TIME.HORIZON()
RISK.YEAR()
ADDITIONAL.YEAR()
TECH.LIST()
TECH.EXCLUDE()
SECTOR.LIST()
OTHER.SECTOR.LIST()
SCENARIO.SOURCES.LIST()
IEA.SCENARIO.LIST()
WEB.REGION.LIST()
SCENARIO.GEOGRAPHIES.LIST()
EQUITY.MARKET.LIST()
GLOBAL.AGGREGATE.SECTOR.LIST()
GLOBAL.AGGREGATE.SCENARIO.SOURCES.LIST()
CREATE.META.PORTFOLIO()
META.INVESTOR.NAME()
META.PORTFOLIO.NAME()
HasBV()
HasRISK()
HasMAP()
HasSB()
}
\seealso{
\link{get_param}

Other functions to get specific 2dii configuration parameters: \code{\link{get_param}}
}
\concept{functions to get specific 2dii configuration parameters}
