% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dropbox_2dii.R
\name{find_dropbox_2dii}
\alias{find_dropbox_2dii}
\alias{path_dropbox_2dii}
\title{Easily access directories in your local copy of 2dii's Dropbox folder}
\usage{
find_dropbox_2dii(...)

path_dropbox_2dii(...)
}
\arguments{
\item{...}{Character vectors, if any values are \code{NA}, the result will also be
\code{NA}.}
}
\value{
A character string.
}
\description{
These functions create cross-platform paths pointing to 2dii's Dropbox
folder:
\itemize{
\item \code{find_dropbox_2dii()} is user-oriented; it checks the output is an existing
path.
\item \code{path_dropbox_2dii()} is low-level, developer-oriented; it doesn't care
if the output is not an existing path.
}
}
\details{
Your projects may need data stored in 2dii's Dropbox folder. Sometimes it is
convenient to have your projects close to the data. But, in this case, it is
a bad idea because the path to 2dii's Dropbox folder has a problematic space
and symbol. (For example, \href{https://github.com/2DegreesInvesting/resources/issues/51}{RStudio's Git pane may not work}.) Instead,
place your projects somewhere with a sane path, such as
\verb{C:/Users/You/git/project/}, and access the data you need with
\code{path_dropbox_2dii()}.
}
\section{Setup for a custom Dropbox folder}{

If the name of your 2dii Dropbox folder is different from the default,
you may add this to .Rprofile (see \code{\link[usethis:edit_r_profile]{usethis::edit_r_profile()}}):\preformatted{options(r2dii_dropbox = "The name of your custom dropbox folder goes here")
}
}

\examples{
# If it points to a nonexistent directory, path_dropbox_2dii() doesn't care
path_dropbox_2dii("path", "to", "nowhere")

# find_dropbox_2dii() points to an existing directory or tells you if not:
# * It is a warning if the Dropbox folder exists but not the nested path
# * It is an error if the Dropbox folder does not exist:
restore <- options(r2dii_dropbox = "No Dropbox folder here")

dropbox_exists()
try(
  find_dropbox_2dii("path", "to", "nowhere")
)

options(restore)
}
\seealso{
\code{\link[=degrees]{degrees()}}, \code{\link[fs:path_home]{fs::path_home()}}.

Other functions to output 2dii paths: 
\code{\link{USER.NAME}()}
}
\concept{functions to output 2dii paths}
