% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{MASTER.DATA.PATH}
\alias{MASTER.DATA.PATH}
\alias{FIN.DATA.PATH}
\alias{CBMARKET.DATA.PATH}
\alias{DATA.STORE.PATH}
\alias{ANALYSIS.INPUTS.PATH}
\title{Create paths to specific 2dii directories, depending on a configuration file}
\usage{
MASTER.DATA.PATH(...)

FIN.DATA.PATH(...)

CBMARKET.DATA.PATH(...)

DATA.STORE.PATH(...)

ANALYSIS.INPUTS.PATH(...)
}
\arguments{
\item{...}{Arguments passed on to \code{fs::path_wd}
\describe{
  \item{ext}{An optional extension to append to the generated path.}
}}
}
\value{
A character string.
}
\description{
The output of this functions depends on the configuration file.
}
\examples{
MASTER.DATA.PATH()

# The output of these functions depends on the configuration file you use
# Defaults to using a configuration file that comes with the package
FIN.DATA.PATH()

# You may use your own configuration file via `options()`
path_to_your_config_file <- example_config("config-toy.yml")
old_options <- options(r2dii.dataraw_config = path_to_your_config_file)
FIN.DATA.PATH()

# Restore
options(old_options)

# Back to default
FIN.DATA.PATH()

CBMARKET.DATA.PATH()
DATA.STORE.PATH()
ANALYSIS.INPUTS.PATH()
}
\seealso{
Replaces \code{dataprep.SetPaths()}.

\link{USER.NAME}
}
\concept{functions to output paths depending on a configuration file}
