% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_dropbox_2dii.R
\name{path_drobox_2dii}
\alias{path_drobox_2dii}
\title{Easily access directories in your local copy of 2dii's Dropbox folder}
\usage{
path_drobox_2dii(...)
}
\arguments{
\item{...}{Arguments passed on to \code{fs::path_wd}
\describe{
  \item{ext}{An optional extension to append to the generated path.}
}}
}
\value{
A character string.
}
\description{
Don't nest your projects in 2dii's Dropbox folder. The path to 2dii's Dropbox
folder contains a space and a symbol which -- among other problems -- may
brake RStudio's Git pane
(\href{https://github.com/2DegreesInvesting/resources/issues/51}{example}). This
function allows you to access files located in a problematic local address
such as \code{C:/Users/You/Dropbox (2{degrees} Investing)/datasets/data.csv} from
a local project located in a sane address such as
\code{C:/Users/You/git/project/}.
}
\examples{
path_drobox_2dii()

path_drobox_2dii("path/to/somewhere")
}
\seealso{
\code{\link[=degrees]{degrees()}}.

Other functions to output 2dii paths: \code{\link{USER.NAME}}
}
\concept{functions to output 2dii paths}
