% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_dropbox_2dii.R
\name{path_dropbox_2dii}
\alias{path_dropbox_2dii}
\title{Easily access directories in your local copy of 2dii's Dropbox folder}
\usage{
path_dropbox_2dii(...)
}
\arguments{
\item{...}{Character vectors, if any values are \code{NA}, the result will also be
\code{NA}.}
}
\value{
A character string.
}
\description{
This function creates cross-platform paths pointing to 2dii's Dropbox folder.
}
\details{
Your projects may need data stored in 2dii's Dropbox folder. Sometimes it is
convenient to have your projects close to the data. But, in this case, it is
a bad idea because the path to 2dii's Dropbox folder has a problematic space
and symbol. (For example, \href{https://github.com/2DegreesInvesting/resources/issues/51}{RStudio's Git pane may not work}.) Instead,
place your projects somewhere with a sane path, such as
\code{C:/Users/You/git/project/}, and access the data you need with
\code{path_dropbox_2dii()}.
}
\examples{
path_dropbox_2dii()

path_dropbox_2dii("path", "to", "somewhere")
}
\seealso{
\code{\link[=degrees]{degrees()}}, \code{\link[fs:path_home]{fs::path_home()}}.

Other functions to output 2dii paths: \code{\link{USER.NAME}}
}
\concept{functions to output 2dii paths}
