% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_examples.R
\name{path_examples}
\alias{path_examples}
\alias{path_example}
\title{Paths to example data installed in any package at \code{.libPaths()}}
\usage{
path_examples(package)

path_example(path, package)
}
\arguments{
\item{package}{Character string giving the name of a single package. \code{NULL}
defaults to the current}

\item{path}{Name of file.}
}
\value{
\itemize{
\item \code{path_example()}: A path to a single example.
\item \code{path_examples()}: A character vector of all built-in examples.
}
}
\description{
Paths to example data installed in any package at \code{.libPaths()}
}
\examples{
path_examples("readr")
path_example("mtcars.csv", package = "readr")
}
\seealso{
Other miscellaneous utility functions: \code{\link{degrees}},
  \code{\link{dropbox_exists}},
  \code{\link{r2dii_packages}}
}
\concept{miscellaneous utility functions}
