% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii_packages.R
\name{r2dii_packages}
\alias{r2dii_packages}
\title{Return a vector giving the name of r2dii packages}
\usage{
r2dii_packages()
}
\value{
String. A vector giving the name of r2dii packages
}
\description{
Return a vector giving the name of r2dii packages
}
\examples{
r2dii_packages()
}
\seealso{
Other miscellaneous utility functions: 
\code{\link{degrees}()},
\code{\link{dropbox_exists}()},
\code{\link{path_examples}()}
}
\concept{miscellaneous utility functions}
