% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_config.R
\name{show_config}
\alias{show_config}
\title{Show a configuration file}
\usage{
show_config(path = get_config(), ...)
}
\arguments{
\item{path}{A string giving the path to a configuration file. \code{NULL} defaults
to \code{\link[=get_config]{get_config()}}.}

\item{...}{Other arguments passed to \code{\link[readr:read_lines]{readr::read_lines()}}, e.g. \code{n_max}.}
}
\value{
Prints to the console the configuration file.
}
\description{
Show a configuration file
}
\examples{
# A configuration file stored in the package
path <- default_config()
show_config(path, n_max = 10)

# Create a new configuration file to show a more general case
config_file <- glue::glue("
  default:
    a_field: a_value
")
config_path <- tempfile()
readr::write_lines(config_file, config_path)

show_config(config_path)
}
\seealso{
\link{FINANCIAL.TIMESTAMP}

Other functions to deal with configuration files: \code{\link{default_config}},
  \code{\link{edit_config}}
}
\concept{functions to deal with configuration files}
