% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{USER.NAME}
\alias{USER.NAME}
\alias{DROPBOX.PATH}
\alias{GIT.PATH}
\alias{WEBTOOL.PATH}
\alias{COMPANY.PATH}
\alias{dbox_port_00}
\alias{SCENARIO.DATA.PATH}
\alias{dbox_port2_10proj}
\alias{EQMARKET.DATA.PATH}
\alias{MARKET.DATA.PATH}
\alias{INDEX.DATA.PATH}
\alias{INDEX.RESULTS.DATA.PATH}
\alias{FUND.DATA.PATH}
\title{Create paths to specific 2dii directories}
\usage{
USER.NAME()

DROPBOX.PATH(...)

GIT.PATH(...)

WEBTOOL.PATH(...)

COMPANY.PATH(...)

dbox_port_00(...)

SCENARIO.DATA.PATH(...)

dbox_port2_10proj(...)

EQMARKET.DATA.PATH(...)

MARKET.DATA.PATH(...)

INDEX.DATA.PATH(...)

INDEX.RESULTS.DATA.PATH(...)

FUND.DATA.PATH(...)
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA.}
}
\value{
A character string.
}
\description{
Create paths to specific 2dii directories
}
\examples{
fs::path_file(fs::path_home())
# Shortcut
USER.NAME()

DROPBOX.PATH()
GIT.PATH()
WEBTOOL.PATH()
COMPANY.PATH()
dbox_port_00("a", "directory")
SCENARIO.DATA.PATH()
dbox_port2_10proj("a", "directory")
EQMARKET.DATA.PATH()
MARKET.DATA.PATH()
INDEX.DATA.PATH()
INDEX.RESULTS.DATA.PATH()
FUND.DATA.PATH()
}
\seealso{
Replaces \code{dataprep.SetPaths()}.

\link{MASTER.DATA.PATH}
}
\concept{functions to output 2dii paths}
