% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclean_names.R
\name{unclean_names}
\alias{unclean_names}
\title{Revert the effect of \code{clean_names_and_groups()}}
\usage{
unclean_names(data, unclean)
}
\arguments{
\item{data}{A dataframe}

\item{unclean}{A dataframe, commonly a version of \code{data} before running
\code{janitor::clean_names(data)}.}
}
\description{
Revert the effect of \code{clean_names_and_groups()}
}
\examples{
library(dplyr)

dirty <- tibble(x.x = 1, y = 1)
names(dirty)

clean <- janitor::clean_names(dirty)
names(clean)

names(unclean_names(clean, dirty))

# Cases where data doesn't change because no name matches unclean names
# exactly -- after applying janitor::clean_names()

unclean <- tibble(y = 1)
data <- tibble(y.y = 1)
data \%>\% unclean_names(unclean)

unclean <- tibble(y.y = 1)
data <- tibble(y = 1)
data \%>\% unclean_names(unclean)
}
\seealso{
Other functions to deal with unclean names: 
\code{\link{clean_names_and_groups}()},
\code{\link{clean_quos}()},
\code{\link{clean_quo}()}
}
\concept{functions to deal with unclean names}
